+++ packages/datadog_flutter_plugin/lib/datadog_internal.dart
@@ -6,5 +6,22 @@
 // extension packages, and is not meant for public use. Anything exposed by this
 // file has the potential to change without notice.
 
+import 'dart:math';
+
 export 'src/attributes.dart';
 export 'src/rum/attributes.dart';
+
+final Random _traceRandom = Random();
+
+String generateTraceId() {
+  // Though traceid is an unsigned 64-bit int, for compatibility
+  // we assume it needs to be a positive signed 64-bit int, so only
+  // use 63-bits.
+  final highBits = _traceRandom.nextInt(1 << 31);
+  final lowBits = BigInt.from(_traceRandom.nextInt(1 << 32));
+
+  var traceId = BigInt.from(highBits) << 32;
+  traceId += lowBits;
+
+  return traceId.toString();
+}